Changed lock function to compare timestamps instead of DateTimeInterface objects directly.
When using optimistic lock with DateTimeInterface based version field a
bug appears due to the use of the === operator for comparing the lock
version and the entity version. This comparison always resolves to false
because the === operator when comparing objects is only true when both
sides are the exact same instance of the object.

To fix the issue I have decided to compare timestamps instead the
DateTimeInterface based objects directly, calling getTimestamp() method
and doing a strict comparison.

Modified OptimisticLockException to use DateTimeInterface instead of
DateTime class.

Fixes #8499